+++ src/php/Compiler/Domain/Emitter/OutputEmitter/NodeEmitter/SetEmitter.php
@@ -18,11 +18,17 @@ final class SetEmitter implements NodeEmitterInterface
     public function emit(AbstractNode $node): void
     {
         assert($node instanceof SetNode);
+        $loc = $node->getStartSourceLocation();
 
-        $this->outputEmitter->emitContextPrefix($node->getEnv(), $node->getStartSourceLocation());
-        $this->outputEmitter->emitStr('\\' . Phel::class . '::set([', $node->getStartSourceLocation());
-        $this->outputEmitter->emitArgList($node->getValues(), $node->getStartSourceLocation());
-        $this->outputEmitter->emitStr('])', $node->getStartSourceLocation());
-        $this->outputEmitter->emitContextSuffix($node->getEnv(), $node->getStartSourceLocation());
+        $this->outputEmitter->emitContextPrefix($node->getEnv(), $loc);
+        $cached = $this->outputEmitter->emitConstantSlotPrefix($node, $loc);
+        $this->outputEmitter->emitStr('\\' . Phel::class . '::set([', $loc);
+        $this->outputEmitter->emitArgList($node->getValues(), $loc);
+        $this->outputEmitter->emitStr('])', $loc);
+        if ($cached) {
+            $this->outputEmitter->emitConstantSlotSuffix($loc);
+        }
+
+        $this->outputEmitter->emitContextSuffix($node->getEnv(), $loc);
     }
 }