+++ src/php/Compiler/Domain/Emitter/OutputEmitter/NodeEmitter/VectorEmitter.php
@@ -17,11 +17,17 @@ final class VectorEmitter implements NodeEmitterInterface
     public function emit(AbstractNode $node): void
     {
         assert($node instanceof VectorNode);
+        $loc = $node->getStartSourceLocation();
 
-        $this->outputEmitter->emitContextPrefix($node->getEnv(), $node->getStartSourceLocation());
-        $this->outputEmitter->emitStr('\Phel::vector([', $node->getStartSourceLocation());
-        $this->outputEmitter->emitArgList($node->getArgs(), $node->getStartSourceLocation());
-        $this->outputEmitter->emitStr('])', $node->getStartSourceLocation());
-        $this->outputEmitter->emitContextSuffix($node->getEnv(), $node->getStartSourceLocation());
+        $this->outputEmitter->emitContextPrefix($node->getEnv(), $loc);
+        $cached = $this->outputEmitter->emitConstantSlotPrefix($node, $loc);
+        $this->outputEmitter->emitStr('\Phel::vector([', $loc);
+        $this->outputEmitter->emitArgList($node->getArgs(), $loc);
+        $this->outputEmitter->emitStr('])', $loc);
+        if ($cached) {
+            $this->outputEmitter->emitConstantSlotSuffix($loc);
+        }
+
+        $this->outputEmitter->emitContextSuffix($node->getEnv(), $loc);
     }
 }